rename subscription-nats to subscription/nats and subscription to subscription/beb
