+++ components/eventing-controller/reconciler/subscription/beb/equality.go
@@ -1,4 +1,4 @@
-package subscription
+package beb
 
 import (
 	eventingv1alpha1 "github.com/kyma-project/kyma/components/eventing-controller/api/v1alpha1"
