+++ components/eventing-controller/reconciler/subscription/beb/event.go
@@ -1,4 +1,4 @@
-package subscription
+package beb
 
 import (
 	corev1 "k8s.io/api/core/v1"
