+++ components/eventing-controller/reconciler/subscription/beb/reconciler_test.go
@@ -1,4 +1,4 @@
-package subscription
+package beb
 
 import (
 	"context"
@@ -1158,8 +1158,8 @@ var _ = BeforeSuite(func(done Done) {
 	useExistingCluster := useExistingCluster
 	testEnv = &envtest.Environment{
 		CRDDirectoryPaths: []string{
-			filepath.Join("../../", "config", "crd", "bases"),
-			filepath.Join("../../", "config", "crd", "external"),
+			filepath.Join("../../../", "config", "crd", "bases"),
+			filepath.Join("../../../", "config", "crd", "external"),
 		},
 		AttachControlPlaneOutput: attachControlPlaneOutput,
 		UseExistingCluster:       &useExistingCluster,
