+++ components/eventing-controller/reconciler/subscription/beb/unit_test.go
@@ -1,4 +1,4 @@
-package subscription
+package beb
 
 import (
 	"testing"
