+++ components/eventing-controller/reconciler/subscription/nats/reconciler_test.go
@@ -1,4 +1,4 @@
-package subscription_nats
+package nats
 
 import (
 	"context"
@@ -381,8 +381,8 @@ var _ = BeforeSuite(func(done Done) {
 
 	testEnv = &envtest.Environment{
 		CRDDirectoryPaths: []string{
-			filepath.Join("../../", "config", "crd", "bases"),
-			filepath.Join("../../", "config", "crd", "external"),
+			filepath.Join("../../../", "config", "crd", "bases"),
+			filepath.Join("../../../", "config", "crd", "external"),
 		},
 		AttachControlPlaneOutput: attachControlPlaneOutput,
 		UseExistingCluster:       &useExistingCluster,
